% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_qc.R
\name{barcode2bitmatrix}
\alias{barcode2bitmatrix}
\title{Convert the barcode vector to a matrix of bit assignments (genes * bits)}
\usage{
barcode2bitmatrix(barcodes)
}
\description{
Convert the barcode vector to a matrix of bit assignments (genes * bits)
}
