% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_qc.R
\name{cellxgene2squarexbit}
\alias{cellxgene2squarexbit}
\title{Convert cell x gene matrix to grid square * bit matrix}
\usage{
cellxgene2squarexbit(counts, grid, genes, barcodes)
}
\description{
Convert cell x gene matrix to grid square * bit matrix
}
