% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_cosmx.R
\name{condenseTissues}
\alias{condenseTissues}
\title{Arrange tissues in xy space to reduce whitespace}
\usage{
condenseTissues(
  xy,
  tissue,
  tissueorder = NULL,
  buffer = 0.2,
  widthheightratio = 8/3
)
}
\arguments{
\item{xy}{A matrix of xy coordinates.}

\item{tissue}{A vector of tissue labels corresponding to the rows of xy.}

\item{tissueorder}{Optional vector specifying the order tissues are tiled in.}

\item{buffer}{Space between tissues.}

\item{widthheightratio}{Desired shape of final xy locations.}
}
\value{
A matrix of updated xy coordinates.
}
\description{
Uses a shelf algorithm: places tallest tissues on the bottom shelf, and so on.
}
