% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_qc.R
\name{FOVEffectsSpatialPlots}
\alias{FOVEffectsSpatialPlots}
\title{Spatial plots of FOV effects:}
\usage{
FOVEffectsSpatialPlots(
  res,
  outdir = NULL,
  bits = "flagged_reportercycles",
  plotwidth = NULL,
  plotheight = NULL
)
}
\description{
Spatial plots of FOV effects:
}
