% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbspatial-package.R
\docType{package}
\name{gbspatial-package}
\alias{gbspatial}
\alias{gbspatial-package}
\title{gbspatial: functions for spatial transcriptomic data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gbarisano/gbspatial}
  \item Report bugs at \url{https://github.com/gbarisano/gbspatial/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Barisano \email{barisano@stanford.edu}

Authors:
\itemize{
  \item Giuseppe Barisano \email{barisano@stanford.edu}
}

}
\keyword{internal}
