% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_qc.R
\name{getNearestNeighborsByFOV}
\alias{getNearestNeighborsByFOV}
\title{Get nearest neighbors, sampling diffusely across other FOVs.}
\usage{
getNearestNeighborsByFOV(x, gridfov, n_neighbors = 10)
}
\description{
Get nearest neighbors, sampling diffusely across other FOVs.
}
