% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_qc.R
\name{mapFlaggedFOVs}
\alias{mapFlaggedFOVs}
\title{Map of which FOVs were flagged:}
\usage{
mapFlaggedFOVs(
  res,
  shownames = TRUE,
  outdir = NULL,
  plotwidth = NULL,
  plotheight = NULL
)
}
\description{
Map of which FOVs were flagged:
}
