% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_spatial_data}
\alias{my_spatial_data}
\title{Example Spatial Coordinates}
\format{
A \code{Seurat} object containing spatial coordinates and expression data.
}
\source{
Downsampled from an in-house dataset
}
\usage{
my_spatial_data
}
\description{
A tiny Seurat object containing spatial transcriptomic data from Nanostring CosMx.
It includes 4 Fields of View (FOVs) across 4 Tissue Microarrays (TMAs).
Useful for testing the package's spatial functions.
}
\details{
Example TMA Spatial Seurat Object

A miniature Seurat object containing spatial transcriptomics data.
}
\keyword{datasets}
