% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_qc.R
\name{run_spatial_qc}
\alias{run_spatial_qc}
\title{Comprehensive Spatial Transcriptomics QC Pipeline}
\usage{
run_spatial_qc(
  counts,
  metadata,
  xy = NULL,
  negcounts = NULL,
  do_nCount = TRUE,
  count_threshold = 20,
  do_cellArea = TRUE,
  area_threshold = 30000,
  do_runfovqc = TRUE,
  panel_name = "Hs_6k",
  max_prop_loss = 0.6,
  max_totalcounts_loss = 0.6,
  do_FOV_boundary = TRUE,
  SplitRatioToLocalThreshold = 0.5,
  do_regional = TRUE,
  bandwidth = 0.01,
  weight_cutoff = 0.08
)
}
\arguments{
\item{counts}{Matrix or data frame of gene expression counts.}

\item{metadata}{Data frame containing cell metadata (must contain 'cell_id').}

\item{xy}{Data frame or matrix of spatial coordinates (required if do_runfovqc or do_regional is TRUE).}

\item{negcounts}{Matrix or data frame of negative probe counts (required if do_regional is TRUE).}

\item{do_nCount}{Logical, perform nCount_RNA QC.}

\item{count_threshold}{Numeric, minimum count threshold.}

\item{do_cellArea}{Logical, perform cell area QC.}

\item{area_threshold}{Numeric, maximum allowed cell area.}

\item{do_runfovqc}{Logical, perform FOV-level QC.}

\item{panel_name}{Character, panel name for barcodes (one of the following: "Hs_IO", "Hs_UCC", "Hs_6k", "Mm_Neuro", "Mm_UCC", "Hs_WTX". Default is "Hs_6k").}

\item{max_prop_loss}{Numeric, FOV QC parameter.}

\item{max_totalcounts_loss}{Numeric, FOV QC parameter.}

\item{do_FOV_boundary}{Logical, perform FOV boundary QC.}

\item{SplitRatioToLocalThreshold}{Numeric, threshold for FOV boundary size ratio.}

\item{do_regional}{Logical, perform regional-level cell QC (SBR).}

\item{bandwidth}{Numeric, bandwidth for spatial smoothing.}

\item{weight_cutoff}{Numeric, cutoff for spatial weighting.}
}
\description{
Comprehensive Spatial Transcriptomics QC Pipeline
}
